research_analytics/requirements.txt
pandas>=2.0.0
pyyaml>=6.0
scikit-learn>=1.3.0
pgmpy>=0.1.25
numpy>=1.24.0
spacy>=3.7.0
